tickets/src/routes/new.ts
import express, { Request, Response } from 'express';
import { body } from 'express-validator';
const router = express.Router();
import { requireAuth, validateRequest } from '@olegs777tickets/common';
import { Ticket } from '../models/ticket';
import { TicketCreatedPublisher } from '../events/publishers/ticket-created-publisher';
import { natsWrapper } from '../nats-wrapper';

router.post(
  '/api/tickets',
  requireAuth,
  [
    body('title').not().isEmpty().withMessage('title is required'),
    body('price')
      .isFloat({ gt: 0 })
      .withMessage('price must be greater then 0'),
  ],
  validateRequest,
  async (req: Request, res: Response) => {
    const { title, price } = req.body;
    const ticket = Ticket.build({
      title,
      price,
      userId: req.user!.id,
    });
    await ticket.save();
    console.log('ticket created', ticket);

    //Publish Event to NATS
    new TicketCreatedPublisher(natsWrapper.client).publish({
      id: ticket.id,
      version: ticket.version,
      title: ticket.title,
      price: ticket.price,
      userId: ticket.userId,
    });

    res.status(201).json(ticket);
  },
);

export { router as createTicketRouter };
